refactor: move `set_raw_external_senders()` to `transaction_context` module
